+++ src/Geno.hpp
@@ -107,7 +107,7 @@ struct variant_block {
   double scale_fac, ac1, mac1, af1, info1, ns1;
   Eigen::ArrayXi ns, ns_case, ns_control, nmales;
   Eigen::ArrayXd af, af_case, af_control, mac, info, cf_burden;
-  Eigen::MatrixXd genocounts;
+  Eigen::MatrixXi genocounts;
   ArrayXb ignored_trait;
   ArrayXb test_fail;
   ArrayXb is_corrected; // for firth/spa
